add subdivision controller, add swagger

+++ src/employee/employee.service.ts
@@ -1,22 +1,44 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { DeleteResult, Repository, UpdateResult } from 'typeorm';
 import { Employee } from '../typeorm/entities/employee.entity';
 import { CreateEmployeeDto } from './dto/create-employee.dto';
+import { Subdivision } from '../typeorm/entities/subdivision.entity';
 
 @Injectable()
 export class EmployeeService {
   constructor(
     @InjectRepository(Employee)
     private usersRepository: Repository<Employee>,
+    @InjectRepository(Subdivision)
+    private subdivisionRepository: Repository<Subdivision>,
   ) {}
 
-  findAll(): Promise<Employee[]> {
-    return this.usersRepository.find();
+  async findAll(): Promise<Employee[]> {
+    const employees: Employee[] = await this.usersRepository.find();
+    const res: Employee[] = [];
+    for (const item of employees) {
+      const subdivisionDto: Subdivision =
+        await this.subdivisionRepository.findOneBy({
+          id: item.subdivisionId,
+        });
+      res.push({ ...item, subdivision: subdivisionDto });
+    }
+    return res;
   }
 
-  findOne(id: number): Promise<Employee> {
-    return this.usersRepository.findOneBy({ id });
+  async findOne(id: number): Promise<Employee> {
+    const employeeDto: Employee = await this.usersRepository.findOneBy({ id });
+
+    employeeDto.subdivisionId = undefined;
+    employeeDto.password = undefined;
+
+    const subdivisionDto: Subdivision =
+      await this.subdivisionRepository.findOneBy({
+        id: employeeDto.subdivisionId,
+      });
+
+    return { ...employeeDto, subdivision: subdivisionDto };
   }
 
   async create(employeeDto: CreateEmployeeDto): Promise<Employee> {
@@ -27,11 +49,28 @@ export class EmployeeService {
     return await this.usersRepository.save(newEmployee);
   }
 
-  async remove(id: string): Promise<void> {
-    await this.usersRepository.delete(id);
+  async remove(id: number): Promise<DeleteResult> {
+    const employee = await this.usersRepository.findOneBy({ id });
+    if (!employee) {
+      throw new HttpException(
+        'Employee not found. Cannot delete employee.',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+    return await this.usersRepository.delete(id);
   }
 
-  async update(id: string, employeeDto: CreateEmployeeDto): Promise<void> {
-    await this.usersRepository.update(id, employeeDto);
+  async update(
+    id: number,
+    employeeDto: CreateEmployeeDto,
+  ): Promise<UpdateResult> {
+    const employee = await this.usersRepository.findOneBy({ id });
+    if (!employee) {
+      throw new HttpException(
+        'Employee not found. Cannot update employee.',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+    return await this.usersRepository.update(id, employeeDto);
   }
 }
